+++ src/spec/interaction.ts
@@ -1,4 +1,5 @@
 import type { TooltipStyleProps } from '@antv/component';
+import { TextStyleProps } from '@antv/g';
 import { BBox, InteractionComponent } from '../runtime';
 import { FisheyeCoordinate } from './coordinateTransform';
 import { TooltipItemValue } from './component';
@@ -25,7 +26,8 @@ export type Interaction =
   | BrushYFilterInteraction
   | BrushXFilterInteraction
   | SliderFilterInteraction
-  | PoptipInteraction;
+  | PoptipInteraction
+  | DrillDownInteraction;
 
 export type InteractionTypes =
   | 'elementHighlight'
@@ -49,6 +51,7 @@ export type InteractionTypes =
   | 'brushYFilter'
   | 'sliderFilter'
   | 'poptip'
+  | 'drillDown'
   | InteractionComponent;
 
 export type BrushHighlightInteraction = {
@@ -231,3 +234,12 @@ export type CustomInteraction = {
   type?: InteractionComponent;
   [key: string]: any;
 };
+
+export type DrillDownInteraction = {
+  type?: 'drillDown';
+  breadCrumb?: {
+    rootText?: string;
+    style?: TextStyleProps;
+    active?: TextStyleProps;
+  };
+} & Record<string, any>;